Fix dependency detection and config

+++ src/connectors/connectoritem.cpp
@@ -490,7 +490,7 @@ ConnectorItem * ConnectorItem::removeConnection(ItemBase * itemBase) {
 			restoreColor(visited);
 			if (itemBase) {
 				DebugDialog::debug(QString("remove from:%1 to:%2 count%3")
-						   .arg((long) this, 0, 16)
+						   .arg((long long) this, 0, 16)
 						   .arg(itemBase->modelPartShared()->title())
 						   .arg(m_connectedTo.count()) );
 			}
@@ -1956,7 +1956,7 @@ void ConnectorItem::debugInfo(const QString & msg)
 	            .arg(this->attachedToViewLayerPlacement())
 	            .arg(this->attachedTo()->wireFlags())
 	            .arg(this->m_hybrid)
-	            .arg((long) this->bus(), 0, 16)
+	            .arg((long long) this->bus(), 0, 16)
 	            .arg(this->m_radius)
 	            .arg(this->m_strokeWidth)
 	            .arg(p.x())
